+++ src/services/axios/get/getSensitivityCalculator/getContinuumData/getContinuumData.tsx
@@ -4,7 +4,7 @@ import {
   DECIMAL_PLACES,
   FREQUENCY_HZ,
   FREQUENCY_MHZ,
-  IW_UNIFORM,
+  IMAGE_WEIGHTING_DEFAULT,
   REFERENCE_COORDINATE_TYPE_GALACTIC,
   REFERENCE_COORDINATE_TYPE_ICRS,
   ROBUST_DEFAULT,
@@ -325,7 +325,8 @@ function GetContinuumData(
     },
     numberOfSubBands: observation?.numSubBands ?? 0,
     spectralAveraging: observation?.spectralAveraging ?? 1,
-    imageWeighting: (dataProductSDP?.data as SDPImageContinuumData)?.weighting ?? IW_UNIFORM,
+    imageWeighting:
+      (dataProductSDP?.data as SDPImageContinuumData)?.weighting ?? IMAGE_WEIGHTING_DEFAULT,
     robust: (dataProductSDP?.data as SDPImageContinuumData)?.robust ?? ROBUST_DEFAULT,
     tapering: (dataProductSDP?.data as SDPImageContinuumData)?.taperValue ?? TAPER_DEFAULT
   };